speaker_filter.py
from pyannote.audio import Pipeline
import torch
from config import TARGET_SPEAKER_EMBEDDING_PATH

# Load diarization pipeline
pipeline = Pipeline.from_pretrained("pyannote/speaker-diarization")

def filter_by_speaker(audio_path):
    diarization = pipeline(audio_path)

    # Load your saved target speaker embedding
    target_embedding = torch.load(TARGET_SPEAKER_EMBEDDING_PATH)

    # Compare embeddings and extract only relevant segments (mocked here)
    print("Filtering speaker segments (requires custom logic)...")
    # Ideally, you’ll extract embeddings per segment and compare using cosine similarity
    return audio_path  # placeholder for now
