requirement.txt
torch
torchaudio
pyaudio
whisper
pyannote.audio
transformers
soundfile
openai-whisper
numpy
sounddevice
